Move proxy entity logic into eventd
